requirements.txt
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.5.4
reportlab==4.1.0
flask>=3.1.0
flask-cors>=6.0.0
